Add initial API documentation
