+++ components/SimulationList.tsx
@@ -3,7 +3,7 @@ import {DateTime} from 'luxon'
 import Link from 'next/link'
 import {useState} from 'react'
 import useSWR from 'swr'
-import {SimulationSummary} from '../lib/db'
+import {SimulationSummary} from '../lib/simulation-types'
 import flagAndName from '../lib/regionEmoji'
 import Clock from '../svg/Clock.svg'
 import simListStyle from './SimulationList.module.css'
