+++ data/openapi-base.json
@@ -1,0 +1,26 @@
+{
+  "openapi": "3.0.0",
+  "info": {
+    "title": "COVID Modeling",
+    "version": "0.0.6",
+    "license": {
+      "name": "MIT License",
+      "url": "https://mit-license.org/"
+    }
+  },
+  "servers": [
+    {
+      "url": "/api"
+    }
+  ],
+  "tags": [
+    {
+      "name": "simulations",
+      "description": "Running simulations and retrieving results"
+    }
+  ],
+  "paths": {},
+  "components": {
+    "schemas": {}
+  }
+}
