+++ lib/crystalcast.ts
@@ -1,5 +1,5 @@
 import {output} from '@covid-policy-modelling/api'
-import {Simulation} from './db'
+import {Simulation} from './simulation-types'
 import {elementSum, extractDiff} from './arrayMath'
 import models from './models'
 // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
