+++ lib/simulation-types.ts
@@ -1,3 +1,5 @@
+import {input, output, RunStatus} from '@covid-policy-modelling/api'
+
 export enum SimulationStatus {
   Complete = 'complete',
   InProgress = 'in-progress',
@@ -50,4 +52,54 @@ export interface Interventions {
  */
 export interface InterventionMap {
   [key: string]: Record<string, Interventions>
+}
+
+export type ModelRun = {
+  model_slug: string
+  status: RunStatus
+  results_data: string | null
+  export_location: string | null
+}
+
+export type Simulation = {
+  id: number
+  region_name: string
+  status: RunStatus
+  subregion_name: string | undefined
+  region_id: string
+  subregion_id: string | undefined
+  github_user_id: number
+  github_user_login: string
+  configuration: input.ModelInput
+  model_runs: ModelRun[]
+  label: string
+  created_at: string
+  updated_at: string
+}
+
+export type SimulationSummary = Omit<Simulation, 'configuration'> & {
+  status: RunStatus
+  configurationSummary: string
+}
+
+type ModelCaseSummary = {
+  cConf: number
+  cHosp: number
+  cDeaths: number
+  peakDeath: input.ISODate
+  peakDailyDeath: number
+}
+
+export type CaseSummary = Record<string, ModelCaseSummary>
+
+/**
+ * The data sent from web UI to backend
+ */
+export interface NewSimulationConfig {
+  regionID: string
+  subregionID?: string
+  label: string
+  r0: number | undefined
+  customCalibrationDate?: ISODate
+  interventionPeriods: input.InterventionPeriod[]
 }
