+++ package.json
@@ -18,6 +18,7 @@
     "@azure/storage-blob": "^12.1.1",
     "@covid-policy-modelling/api": "^0.10.1",
     "@hapi/joi": "^17.1.1",
+    "@types/swagger-ui-react": "^3.35.2",
     "axios": "^0.21.1",
     "classnames": "^2.2.6",
     "cleave.js": "^1.5.10",
@@ -40,10 +41,12 @@
     "serverless-mysql": "^1.5.3",
     "source-map-support": "^0.5.16",
     "sql-template-strings": "^2.2.2",
+    "swagger-ui-react": "^4.0.0-beta.2",
     "swr": "^0.2.0"
   },
   "devDependencies": {
     "@fullhuman/postcss-purgecss": "^2.1.0",
+    "@openapi-contrib/json-schema-to-openapi-schema": "^2.1.0",
     "@types/classnames": "^2.2.10",
     "@types/cleave.js": "^1.4.1",
     "@types/cookie": "^0.3.3",
@@ -88,6 +91,7 @@
     "prettier": "^1.19.1",
     "prompts": "^2.3.2",
     "source-map": "^0.7.3",
+    "swagger-inline": "^4.1.5",
     "tailwindcss": "^1.2.0",
     "ts-json-schema-generator": "^0.67.1",
     "ts-node": "^8.10.1",
