+++ pages/simulations/index.tsx
@@ -4,7 +4,7 @@ import {useRouter} from 'next/router'
 import AppFrame from '../../components/AppFrame'
 import button from '../../components/styles/button.module.css'
 import * as db from '../../lib/db'
-import {SimulationSummary} from '../../lib/db'
+import {SimulationSummary} from '../../lib/simulation-types'
 import handleError from '../../lib/handle-error'
 import {withDB} from '../../lib/mysql'
 import {ensureSession} from '../../lib/session'
